+++ src/deckgl/raster-mesh-layer/raster-mesh-layer.js
@@ -185,9 +185,9 @@ export default class RasterMeshLayer extends SimpleMeshLayer {
     if (this.state.images) {
       for (const image of Object.values(this.state.images)) {
         if (Array.isArray(image)) {
-          image.map((x) => x.delete());
+          image.map((x) => x && x.delete());
         } else {
-          image.delete();
+          image && image.delete();
         }
       }
     }
